Migrate scenarios to DAML Scripts
Now that DAML Script runs in DAML Studio, we are starting to migrate
our examples over to DAML Script before we deprecate scenarios.

Most of the stuff in this PR is very straightforward. `fetch` is not a
ledger API command so we use `queryContractId` instead (which is
equivalent to the ACS endpoint + client-side filtering).

I don’t know your version policy but you might want to hold off on
merging this until 1.5.0 is released.